requirements.txt
streamlit
langchain
langchain-community
langchain-groq
pdfplumber
pytesseract
pillow
pandas
openpyxl
python-dateutil
regex
langchain_core